NXP-26710: move RequestContext and RequestContextFilter to web-common
