+++ nuxeo-features/nuxeo-automation/nuxeo-automation-server/src/main/java/org/nuxeo/ecm/automation/server/jaxrs/batch/JSONBatchBlobDecoder.java
@@ -22,7 +22,7 @@ package org.nuxeo.ecm.automation.server.jaxrs.batch;
 
 import org.nuxeo.ecm.automation.core.util.JSONBlobDecoder;
 import org.nuxeo.ecm.core.api.Blob;
-import org.nuxeo.ecm.webengine.jaxrs.context.RequestContext;
+import org.nuxeo.ecm.platform.web.common.RequestContext;
 import org.nuxeo.runtime.api.Framework;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
